requirements.txt
pyahocorasick==2.0.0
rapidfuzz==3.6.1
Unidecode==1.3.7